pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brax"
version = "0.13.0"
description = "A differentiable physics engine written in JAX."
authors = [{ name = "Brax Authors", email = "no-reply@google.com" }]
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = ["JAX", "reinforcement learning", "rigidbody", "physics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "absl-py",
    "dataclasses; python_version < '3.7'",
    "etils",
    "flask",
    "flask_cors",
    "flax",
    "jax>=0.4.6",
    "jaxlib>=0.4.6",
    "jaxopt",
    "jinja2",
    "ml_collections",
    "mujoco",
    "mujoco-mjx",
    "numpy",
    "optax",
    "orbax-checkpoint>=0.11.22",
    "Pillow",
    "scipy",
    "tensorboardX",
    "trimesh",
    "typing-extensions",
]

[project.optional-dependencies]
develop = ["pytest", "transforms3d", "gym", "dm_env"]

[project.urls]
Homepage = "http://github.com/google/brax"

[tool.hatch.build.targets.wheel]
packages = ["brax"]
exclude = [
    "/datasets",
    "/docs",
    "/notebooks",
    "/tests",
    "brax/experimental/barkour/assets",
    "brax/experimental/barkour/data",
]

[tool.hatch.build.targets.sdist]
exclude = [
    "/datasets",
    "/docs",
    "/notebooks",
    "brax/experimental/barkour/assets",
    "brax/experimental/barkour/data",
]

[tool.isort]
force_single_line = true
force_sort_within_sections = true
lexicographical = true
single_line_exclusions = ["typing"]
order_by_type = false
group_by_package = true
line_length = 120
use_parentheses = true
multi_line_output = 3
skip_glob = ["**/*.ipynb"]

[tool.pyink]
line-length = 80
unstable = true
pyink-indentation = 2
pyink-use-majority-quotes = true
extend-exclude = '''(
 .ipynb$
)'''

[tool.ruff]
line-length = 80
exclude = ["*.ipynb"]
indent-width = 2

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
line-ending = "auto"

[tool.ruff.lint]
# Enable pycodestyle, pyflakes, and isort rules
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
# Translate isort settings to ruff
force-single-line = true
force-sort-within-sections = true
single-line-exclusions = ["typing"]
order-by-type = false
lines-after-imports = 2
# Note: ruff doesn't have exact equivalents for:
# - lexicographical (ruff uses natural sort by default)
# - group_by_package (not directly supported)
# - use_parentheses (ruff handles this automatically)
# - multi_line_output (ruff uses its own style)
